mesh-599: resolved reviews
