+++ repository/src/main/java/org/apache/atlas/repository/store/graph/v2/preprocessor/resource/AbstractResourcePreProcessor.java
@@ -29,7 +29,6 @@ import org.apache.atlas.repository.graphdb.AtlasEdgeDirection;
 import org.apache.atlas.repository.graphdb.AtlasVertex;
 import org.apache.atlas.repository.store.graph.v2.EntityGraphRetriever;
 import org.apache.atlas.repository.store.graph.v2.preprocessor.PreProcessor;
-import org.apache.atlas.repository.store.graph.v2.preprocessor.PreProcessorUtils;
 import org.apache.atlas.type.AtlasTypeRegistry;
 import org.apache.atlas.utils.AtlasPerfMetrics;
 import org.slf4j.Logger;
@@ -37,9 +36,11 @@ import org.slf4j.LoggerFactory;
 
 import java.util.Iterator;
 
-import static org.apache.atlas.repository.Constants.ACTIVE_STATE_VALUE;
-import static org.apache.atlas.repository.Constants.ASSET_RELATION_ATTR;
-import static org.apache.atlas.repository.Constants.STATE_PROPERTY_KEY;
+import static org.apache.atlas.AtlasErrorCode.OPERATION_NOT_SUPPORTED;
+import static org.apache.atlas.repository.Constants.*;
+import static org.apache.atlas.repository.Constants.TYPE_NAME_PROPERTY_KEY;
+import static org.apache.atlas.repository.store.graph.v2.preprocessor.PreProcessorUtils.DAAP_ARCHIVED_STATUS;
+import static org.apache.atlas.repository.store.graph.v2.preprocessor.PreProcessorUtils.DAAP_STATUS_ATTR;
 
 public abstract class AbstractResourcePreProcessor implements PreProcessor {
     private static final Logger LOG = LoggerFactory.getLogger(AbstractResourcePreProcessor.class);
@@ -64,7 +65,7 @@ public abstract class AbstractResourcePreProcessor implements PreProcessor {
                 //Found linked asset in payload
                 AtlasVertex assetVertex = entityRetriever.getEntityVertex(asset);
 
-                PreProcessorUtils.validateProductStatus(assetVertex);
+                validateProductStatus(assetVertex);
 
                 assetEntity = entityRetriever.toAtlasEntityHeaderWithClassifications(assetVertex);
 
@@ -125,6 +126,19 @@ public abstract class AbstractResourcePreProcessor implements PreProcessor {
         }
 
         return ret;
+    }
+
+    public static void validateProductStatus(AtlasVertex assetVertex) throws AtlasBaseException {
+        if (assetVertex != null) {
+            if (DATA_PRODUCT_ENTITY_TYPE.equals(assetVertex.getProperty(TYPE_NAME_PROPERTY_KEY, String.class))) {
+                String entityState = assetVertex.getProperty(STATE_PROPERTY_KEY, String.class);
+                String daapStatus = assetVertex.getProperty(DAAP_STATUS_ATTR, String.class);
+
+                if ((AtlasEntity.Status.DELETED.name().equals(entityState)) && DAAP_ARCHIVED_STATUS.equals(daapStatus)) {
+                    throw new AtlasBaseException(OPERATION_NOT_SUPPORTED, "Cannot update DataProduct that is Archived!");
+                }
+            }
+        }
     }
 
     private void verifyAssetAccess(AtlasEntityHeader asset, AtlasPrivilege assetPrivilege,
